+++ crates/ruma-client-api/src/room.rs
@@ -4,6 +4,8 @@ pub mod aliases;
 pub mod create_room;
 pub mod get_event_by_timestamp;
 pub mod get_room_event;
+#[cfg(feature = "unstable-msc3266")]
+pub mod get_summary;
 pub mod report_content;
 pub mod upgrade_room;